Prevent uncessary listeners to the oplog changes
