+++ app/metrics/server/index.js
@@ -1,6 +1,7 @@
 import { metrics } from './lib/metrics';
 import StatsTracker from './lib/statsTracker';
 
+import './lib/collectMetrics';
 import './callbacksMetrics';
 
 export {
