--- app/models/server/oplogEvents.js
@@ -1,3 +1,0 @@
-import { EventEmitter } from 'events';
-
-export const oplogEvents = new EventEmitter();
